+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * ADD: ignore for git
 * ADD: ignore for style
 * ADD: README
+* IMP: ignore get if statements
 
 ## 1.0.0 - 28-03-19
 * Initial Setup